implemented delete_student from model and integrated it within the student panel

+++ app/backend/src/Controller/student_controller.py
@@ -34,10 +34,18 @@ class StudentController:
         if self.user_role != 'admin':
             messagebox.showerror("Access Denied!", "You don't have permission to delete user!")
             return
+        
+        confirm = messagebox.askyesno("Confirm Delete", f"Are you sure you want to delete student {student_id}?")
+        if not confirm:
+            return
+        
         success = self.model.delete_student(student_id)
         if success:
             messagebox.showinfo("Success", "Student deleted successfully")
             self.views.populate_students()
+
+        else:
+            messagebox.showerror("Error", "Student not found.")
 
     def search_student(self):
         pass
